src/requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
asyncpg==0.30.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
